+++ api/typescript/src/engine/lib/kurtosis_context/generic_engine_client.ts
@@ -7,7 +7,9 @@ import {
     DestroyEnclaveArgs,
     GetEnclavesResponse,
     GetEngineInfoResponse,
-    StopEnclaveArgs
+    StopEnclaveArgs,
+    GetUserServiceLogsArgs,
+    GetUserServiceLogsResponse,
 } from "../../kurtosis_engine_rpc_api_bindings/engine_service_pb";
 
 export interface GenericEngineClient {
@@ -17,4 +19,5 @@ export interface GenericEngineClient {
     stopEnclave(stopEnclaveArgs: StopEnclaveArgs): Promise<Result<null, Error>>
     destroyEnclave(destroyEnclaveArgs: DestroyEnclaveArgs): Promise<Result<null, Error>>
     clean(cleanArgs: CleanArgs): Promise<Result<CleanResponse, Error>>
+    getUserServiceLogs(getUserServiceLogsArgs: GetUserServiceLogsArgs): Promise<Result<GetUserServiceLogsResponse, Error>>
 }