+++ api/typescript/src/engine/lib/kurtosis_context/grpc_node_engine_client.ts
@@ -11,7 +11,9 @@ import type {
     DestroyEnclaveArgs,
     GetEnclavesResponse,
     GetEngineInfoResponse,
-    StopEnclaveArgs
+    StopEnclaveArgs,
+    GetUserServiceLogsArgs,
+    GetUserServiceLogsResponse
 } from "../../kurtosis_engine_rpc_api_bindings/engine_service_pb";
 
 export class GrpcNodeEngineClient implements GenericEngineClient {
@@ -134,7 +136,7 @@ export class GrpcNodeEngineClient implements GenericEngineClient {
 
         const cleanResult: Result<CleanResponse, Error> = await cleanPromise;
         if (cleanResult.isErr()) {
-            return err(cleanResult.error)
+            return err(cleanResult.error);
         }
 
         const cleanResponse: CleanResponse = cleanResult.value;
@@ -158,10 +160,33 @@ export class GrpcNodeEngineClient implements GenericEngineClient {
         });
         
         const getEnclavesResponseResult: Result<GetEnclavesResponse, Error> = await getEnclavesPromise;
-        if (!getEnclavesResponseResult.isOk()) {
-            return err(getEnclavesResponseResult.error)
+        if (getEnclavesResponseResult.isErr()) {
+            return err(getEnclavesResponseResult.error);
         }
 
         return ok(getEnclavesResponseResult.value);
+    }
+
+    public async getUserServiceLogs(getUserServiceLogsArgs: GetUserServiceLogsArgs): Promise<Result<GetUserServiceLogsResponse, Error>> {
+        const getUserServiceLogsPromise: Promise<Result<GetUserServiceLogsResponse, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.getUserServiceLogs(getUserServiceLogsArgs, (error: ServiceError | null, response?: GetUserServiceLogsResponse) => {
+                if (error === null) {
+                    if (!response) {
+                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")))
+                    } else {
+                        resolve(ok(response));
+                    }
+                } else {
+                    resolve(err(error))
+                }
+            })
+        })
+
+        const getUserServiceLogsResponseResult: Result<GetUserServiceLogsResponse, Error> = await getUserServiceLogsPromise;
+        if (getUserServiceLogsResponseResult.isErr()) {
+            return err(getUserServiceLogsResponseResult.error);
+        }
+
+        return ok(getUserServiceLogsResponseResult.value);
     }
 }